Rename dictionary validators
